Add files via upload
HubStreet1: Moved Marauder Shields stuntactor to make room for additions in HubStreet2. Note: Marauder Shields POI is still in original position/location.

HubStreet2: First touch of the Terminus Mercenaries additions. Placed three YMIRs and five LOKIs to the nearest truck just opposite of the barricade. Also placed the Eclipse Salarian near one of the mechs, with the idea that maybe he's tinkering with one of them? No other stuntactors have been touched yet, so as to give Lonely Wolf as much freedom as possible with the Volus. Will resume with the other mercs when Volus are finished.