Add files via upload
HubStreet1: Moved three Omega mercenaries around, one of them needs a smoking animation that I did not have access to, the other two are by the start of the level conversing. The N7 engineer and destroyer have also been moved to around the start of the level, chatting opposite on the opposite side of the mechs.